+++ src/components/common/Button/style.css
@@ -1,0 +1,10 @@
+.button-basic {
+  border: none;
+  background-color: transparent;
+}
+
+.button-circle {
+  border-radius: 50%;
+  aspect-ratio: 1 / 1;
+  background-color: #e5e5e5;
+}
