+++ src/styles/button.css
@@ -1,8 +1,8 @@
 .button-box {
   width: 100%;
   text-align: right;
-}
 
-.button-text {
-  margin-right: 10px;
+  position: absolute;
+  bottom: 25px;
+  right: 33px;
 }
